Move PendingWriteQueueDrainer functionality to IndexingPendingWriteQueue
